// web_assets/kone-design-system-main/packages/kone-ds-webcomponents/src/components/atoms/kfds-button/kfds-button.enum.ts
export enum ButtonStyle {
    Primary = 'primary',
    Secondary = 'secondary',
    Tertiary = 'tertiary',
    Destructive = 'destructive',
    Borderless = 'borderless'
};

export enum ButtonType {
    Button = 'button',
    Submit = 'submit',
    Reset = 'reset'
};

export enum ButtonSize {
    Extra = 'extra',
    Big = 'big',
    Small = 'small',
    Medium = 'medium'
};